python/requirements.txt
Flask
pipwin
SpeechRecognition
nltk
pyaudio
pymongo
python-dotenv